Lower per-chunk inference host-RAM plateau
Two changes targeting the ~52% RAM plateau on 32 GB inference boxes, with
the goal of fitting on a half-size box (same GPU) at <=45%:

- Disable cuDNN benchmark mode. With variable bucket shapes (per-bucket
  sequence lengths + partial final sub-batches) the autotuner re-searches
  constantly for little gain and inflates the host-side cuDNN footprint by
  trial-loading multiple algorithms and holding the largest workspace.
- Stop widening the S2 validity mask to int32. It is bool out of
  _load_scl_mask and stays resident in ChunkData for the whole chunk;
  consumers (np.nonzero, .sum) handle bool directly, so this is a free 4x
  shrink on a resident buffer. Also drops the per-batch int32 cast.

Docstrings and test fixtures updated to the bool mask dtype.

+++ src/tessera_embeddings/inference/dataset.py
@@ -199,7 +199,7 @@ class MosaicChunkInferenceDataset:
 
         # S2: (T_s2, H, W, 10) → (T_s2, B, 10) → (B, T_s2, 10)
         s2_bands = self._s2_bands[:, rows, cols, :].transpose(1, 0, 2)
-        s2_masks = self._s2_masks[:, rows, cols].T.astype(np.int32, copy=False)
+        s2_masks = self._s2_masks[:, rows, cols].T
         s2_doys = np.broadcast_to(self._s2_doys[None, :], (b, len(self._s2_doys))).astype(np.int32, copy=False)
 
         s1_asc_bands = self._s1_asc_bands[:, rows, cols, :].transpose(1, 0, 2)
